Decouple the extension definitions from their runtime implementations
